requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cryptography==2.6.1
DateTime==4.3
entrypoints==0.3
freezegun==0.3.15
google==2.0.3
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-spanner==1.17.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
grpcio-gcp==0.2.2
httplib2==0.18.1
idna==2.10
keyring==17.1.1
keyrings.alt==3.1.1
mock==4.0.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
PyGObject==3.30.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.26
requests==2.24.0
retrying==1.3.3
rsa==4.6
SecretStorage==2.3.1
six==1.12.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.9
zope.interface==5.1.0
